fix: rename type files for no using additional path
