+++ packages/babyjubjub/src/index.ts
@@ -1,5 +1,4 @@
-/* eslint-disable import/no-useless-path-segments */
-export { F } from '../src/types/ff'
+export { F } from './types/ff'
 export { Fp } from './fp'
 export { Fr } from './fr'
 export { Point } from './point'
